# doc/specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Store Daily QA System API",
    "description": "API for store staff daily QA system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/insights/daily": {
      "get": {
        "tags": ["insights"],
        "summary": "Get Daily Insights",
        "description": "日次Insights取得\n\n指定された条件に一致する日次Insightsを取得します。",
        "operationId": "get_daily_insights_api_insights_daily_get",
        "parameters": [
          {
            "name": "report_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "レポート対象日",
              "title": "Report Date"
            },
            "description": "レポート対象日"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily insights retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InsightResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "404": {
            "description": "Insights not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/insights/daily/generate": {
      "post": {
        "tags": ["insights"],
        "summary": "Generate Daily Insights",
        "description": "日次Insights生成\n\n指定された条件で新しい日次Insightsを生成します。",
        "operationId": "generate_daily_insights_api_insights_daily_generate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsightGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Daily insights generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightGenerateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/insights/daily/{insight_id}": {
      "put": {
        "tags": ["insights"],
        "summary": "Save Daily Insights",
        "description": "日次Insights保存・更新\n\n指定されたInsightsの内容を保存・更新します。",
        "operationId": "save_daily_insights_api_insights_daily__insight_id__put",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Insights ID",
              "title": "Insight Id"
            },
            "description": "Insights ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InsightSaveRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Daily insights saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "404": {
            "description": "Insights not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      },
      "get": {
        "tags": ["insights"],
        "summary": "Get Insight Detail",
        "description": "Insights詳細取得\n\n指定されたIDのInsights詳細情報を取得します。",
        "operationId": "get_insight_detail_api_insights_daily__insight_id__get",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Insights ID",
              "title": "Insight Id"
            },
            "description": "Insights ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Insight detail retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InsightResponse" }
              }
            }
          },
          "404": {
            "description": "Insights not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/insights/daily/{insight_id}/feedback": {
      "put": {
        "tags": ["insights"],
        "summary": "Save Feedback",
        "description": "フィードバック保存\n\n指定されたInsightsにフィードバックを保存します。",
        "operationId": "save_feedback_api_insights_daily__insight_id__feedback_put",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Insights ID",
              "title": "Insight Id"
            },
            "description": "Insights ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsightFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feedback saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "404": {
            "description": "Insights not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/insights/daily/history": {
      "get": {
        "tags": ["insights"],
        "summary": "Get Insights History",
        "description": "Insights履歴取得\n\nInsightsの履歴を取得します。",
        "operationId": "get_insights_history_api_insights_daily_history_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "ページ番号",
              "default": 1,
              "title": "Page"
            },
            "description": "ページ番号"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "1ページあたりの件数",
              "default": 30,
              "title": "Limit"
            },
            "description": "1ページあたりの件数"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "responses": {
          "200": {
            "description": "Insights history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorResponse" }
        },
        "type": "object",
        "required": ["error"],
        "title": "APIError",
        "description": "API エラーレスポンス"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "エラーコード"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "エラーメッセージ"
          },
          "details": {
            "anyOf": [{}, { "type": "null" }],
            "title": "Details",
            "description": "エラー詳細"
          }
        },
        "type": "object",
        "required": ["code", "message"],
        "title": "ErrorResponse",
        "description": "エラーレスポンス"
      },
      "InsightFeedbackRequest": {
        "properties": {
          "user_feedback": {
            "type": "string",
            "title": "User Feedback",
            "description": "ユーザーフィードバック"
          }
        },
        "type": "object",
        "required": ["user_feedback"],
        "title": "InsightFeedbackRequest",
        "description": "フィードバック保存リクエスト"
      },
      "InsightGenerateRequest": {
        "properties": {
          "agent_type": {
            "type": "string",
            "minLength": 1,
            "title": "Agent Type",
            "description": "エージェントタイプ",
            "default": "daily_summary"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt",
            "description": "生成指示"
          },
          "session_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Session Id",
            "description": "セッションID"
          },
          "report_date": {
            "type": "string",
            "format": "date",
            "title": "Report Date",
            "description": "レポート対象日"
          }
        },
        "type": "object",
        "required": ["agent_type", "prompt", "session_id", "report_date"],
        "title": "InsightGenerateRequest",
        "description": "日次Insights生成リクエスト"
      },
      "InsightGenerateResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response",
            "description": "生成されたInsightsテキスト"
          },
          "agent_type": {
            "type": "string",
            "title": "Agent Type",
            "description": "エージェントタイプ"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "セッションID"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id",
            "description": "アクターID"
          },
          "summary_id": {
            "type": "string",
            "title": "Summary Id",
            "description": "サマリーID"
          }
        },
        "type": "object",
        "required": [
          "response",
          "agent_type",
          "session_id",
          "actor_id",
          "summary_id"
        ],
        "title": "InsightGenerateResponse",
        "description": "Insights生成レスポンス"
      },
      "InsightHistoryItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Insights ID"
          },
          "report_date": {
            "type": "string",
            "title": "Report Date",
            "description": "レポート対象日"
          },
          "user_cd": {
            "type": "string",
            "title": "User Cd",
            "description": "ユーザーコード"
          },
          "str_cd": {
            "type": "string",
            "title": "Str Cd",
            "description": "店舗コード"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新日時"
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_date",
          "user_cd",
          "str_cd",
          "created_at",
          "updated_at"
        ],
        "title": "InsightHistoryItem",
        "description": "Insights履歴項目"
      },
      "InsightHistoryResponse": {
        "properties": {
          "insights": {
            "items": { "$ref": "#/components/schemas/InsightHistoryItem" },
            "type": "array",
            "title": "Insights",
            "description": "Insights一覧"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "総件数"
          },
          "current_page": {
            "type": "integer",
            "title": "Current Page",
            "description": "現在のページ"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "総ページ数"
          }
        },
        "type": "object",
        "required": ["insights", "total_count", "current_page", "total_pages"],
        "title": "InsightHistoryResponse",
        "description": "Insights履歴レスポンス"
      },
      "InsightResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Insights ID"
          },
          "user_cd": {
            "type": "string",
            "title": "User Cd",
            "description": "ユーザーコード"
          },
          "str_cd": {
            "type": "string",
            "title": "Str Cd",
            "description": "店舗コード"
          },
          "report_date": {
            "type": "string",
            "format": "date-time",
            "title": "Report Date",
            "description": "レポート対象日"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "ステータス"
          },
          "report_text": {
            "type": "string",
            "title": "Report Text",
            "description": "レポートテキスト"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "会話セッションID"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id",
            "description": "ユーザー/アクターID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新日時"
          },
          "user_feedback": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Feedback",
            "description": "ユーザーフィードバック"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_cd",
          "str_cd",
          "report_date",
          "status",
          "report_text",
          "session_id",
          "actor_id",
          "created_at",
          "updated_at"
        ],
        "title": "InsightResponse",
        "description": "日次Insightsレスポンス"
      },
      "InsightSaveRequest": {
        "properties": {
          "report_date": {
            "type": "string",
            "format": "date-time",
            "title": "Report Date",
            "description": "レポート対象日"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "セッション状態"
          },
          "report_text": {
            "type": "string",
            "title": "Report Text",
            "description": "レポートテキスト"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "会話セッションID"
          }
        },
        "type": "object",
        "required": ["report_date", "status", "report_text", "session_id"],
        "title": "InsightSaveRequest",
        "description": "日次Insights保存リクエスト"
      },
      "InsightUpdateResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "更新結果メッセージ"
          }
        },
        "type": "object",
        "required": ["message"],
        "title": "InsightUpdateResponse",
        "description": "Insights更新レスポンス"
      },
...
    }
  }
}